Fix GCC warnings and resolve double tooltip UI bug.
- Fixed signedness comparison warning in sys_mmap.
- Removed unused variables from OTA and initrd.
- Updated ORCHESTRATOR_REPORT to reflect a clean build.
- Fixed double tooltips in web_ui by swapping 'title' for 'data-title' dynamically on hover, preserving attribute for Playwright verification tests.

+++ kernel/arch/x86_64/syscall.c
@@ -602,7 +602,7 @@ static int64_t sys_mmap(void* addr, size_t len, int prot, int flags, int fd, int
             }
         } else if (is_shmfs && shm_obj) {
             /* Check prot mask for Ashmem compatibility */
-            if ((prot & shm_obj->prot_mask) != prot) {
+            if (((uint32_t)prot & shm_obj->prot_mask) != (uint32_t)prot) {
                 return -EPERM;
             }
